+++ src/Taxes/Models/TaxRate.php
@@ -19,6 +19,7 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Collection;
+use Konekt\Address\Concerns\Zoneable;
 use Konekt\Address\Contracts\Zone;
 use Konekt\Address\Models\ZoneProxy;
 use Vanilo\Contracts\Schematized;
@@ -45,11 +46,8 @@ use Vanilo\Taxes\Traits\BelongsToTaxCategory;
  * @property Carbon $created_at
  * @property Carbon $updated_at
  * @property-read TaxCategory|null $taxCategory
- * @property-read Zone|null $zone
  *
  * @method static Builder byTaxCategory(TaxCategory|int $taxCategory)
- * @method Builder forZone(Zone|int $zone)
- * @method Builder forZones(array|Collection $zones)
  * @method Builder actives()
  * @method Builder inactives()
  */
@@ -57,6 +55,7 @@ class TaxRate extends Model implements TaxRateContract
 {
     use BelongsToTaxCategory;
     use ConfigurableModel;
+    use Zoneable;
 
     protected $guarded = ['id', 'created_at', 'updated_at'];
 
@@ -97,11 +96,6 @@ class TaxRate extends Model implements TaxRateContract
         return $this->name;
     }
 
-    public function getZone(): ?Zone
-    {
-        return $this->zone;
-    }
-
     public function getTaxCategory(): ?TaxCategory
     {
         return $this->taxCategory;
@@ -126,11 +120,6 @@ class TaxRate extends Model implements TaxRateContract
         return SchemaDefinition::wrap($this->getCalculator());
     }
 
-    public function zone(): BelongsTo
-    {
-        return $this->belongsTo(ZoneProxy::modelClass(), 'zone_id', 'id');
-    }
-
     public function scopeActives(Builder $query): Builder
     {
         return $query->where('is_active', true);
@@ -139,22 +128,6 @@ class TaxRate extends Model implements TaxRateContract
     public function scopeInactives(Builder $query): Builder
     {
         return $query->where('is_active', false);
-    }
-
-    public function scopeForZone(Builder $query, Zone|int $zone): Builder
-    {
-        return $query->where('zone_id', is_int($zone) ? $zone : $zone->id);
-    }
-
-    public function scopeForZones(Builder $query, array|Collection $zones): Builder
-    {
-        if (is_array($zones)) {
-            $zones = array_map(fn (Zone|int $zone) => is_int($zone) ? $zone : $zone->id, $zones);
-        } else {
-            $zones = $zones->map(fn (Zone $zone) => $zone->id);
-        }
-
-        return $query->whereIn('zone_id', $zones);
     }
 
     public function scopeByTaxCategory(Builder $query, TaxCategory|int $taxCategory): Builder